requirements.txt
altgraph==0.17.2
apturl==0.5.2
asgiref==3.4.1
auto-py-to-exe==2.19.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
chromedriver-autoinstaller==0.2.2
Click==7.0
colorama==0.4.3
command-not-found==0.3
configparser==5.0.2
crayons==0.4.0
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
duplicity==0.8.12.0
Eel==0.12.4
entrypoints==0.3
fasteners==0.14.1
fonttools==4.33.2
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
httplib2==0.14.0
idna==2.8
joblib==1.1.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.1
monotonic==1.5
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.0
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
packaging==21.3
pandas==1.4.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
pydotplus==2.0.2
PyGObject==3.36.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==3.0.8
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
schedule==1.1.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sqlparse==0.4.2
systemd-python==234
threadpoolctl==3.1.0
toml==0.10.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
webdriver-manager==3.4.2
whichcraft==0.6.1
whitenoise==6.0.0
xkit==0.0.0
zope.event==4.5.0
zope.interface==5.4.0
